requirements.txt
# Core
aiogram==3.0.0
sqlalchemy==2.0.0
alembic==1.12.0
asyncpg==0.29.0

# Scheduler & Background Tasks
apscheduler==3.10.0
celery==5.3.0
redis==5.0.0

# APIs & Integrations
google-api-python-client==2.80.0
google-auth-oauthlib==1.0.0
msal==1.24.0
pyjwt[crypto]==2.8.0
requests==2.31.0
python-zoom==0.1.0

# AI & Machine Learning
openai==0.28.0
langchain==0.0.340
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.3.0

# Web & API
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0
httpx==0.25.0
python-multipart==0.0.6

# Security
cryptography==41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
reportlab==4.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
cachetools==5.3.0
colorama==0.4.6

# Monitoring & Logging
sentry-sdk==1.35.0
structlog==23.2.0
prometheus-client==0.18.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
httpx==0.25.0

# Development
black==23.9.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.0
pre-commit==3.4.0